+++ src/main/java/net/sf/jabref/gui/groups/RemoveFromGroupAction.java
@@ -2,7 +2,6 @@ package net.sf.jabref.gui.groups;
 
 import java.awt.event.ActionEvent;
 import java.util.List;
-import java.util.Optional;
 
 import javax.swing.AbstractAction;
