+++ src/main/java/net/sf/jabref/model/groups/AbstractGroup.java
@@ -1,11 +1,8 @@
 package net.sf.jabref.model.groups;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 
-import net.sf.jabref.model.database.BibDatabase;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.search.SearchMatcher;
