+++ src/main/java/net/sf/jabref/model/groups/GroupEntryChanger.java
@@ -2,7 +2,6 @@ package net.sf.jabref.model.groups;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
 
 import net.sf.jabref.model.FieldChange;
 import net.sf.jabref.model.entry.BibEntry;
