+++ src/main/java/net/sf/jabref/model/groups/GroupHierarchyType.java
@@ -1,7 +1,5 @@
 package net.sf.jabref.model.groups;
 
-import java.util.Optional;
-
 public enum GroupHierarchyType {
 
     /**
