+++ src/main/java/net/sf/jabref/model/groups/GroupTreeNode.java
@@ -3,7 +3,6 @@ package net.sf.jabref.model.groups;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 import net.sf.jabref.model.FieldChange;
