+++ src/main/java/net/sf/jabref/shared/listener/OracleNotificationListener.java
