+++ src/test/java/net/sf/jabref/logic/bibtex/FieldContentParserTest.java
