+++ src/test/java/net/sf/jabref/logic/bibtex/comparator/EntryComparatorTest.java
