+++ src/test/java/net/sf/jabref/logic/cleanup/MoveFilesCleanupTest.java
