+++ src/test/java/net/sf/jabref/logic/layout/format/HTMLCharsTest.java
