+++ src/test/java/net/sf/jabref/logic/layout/format/RemoveBracketsAddCommaTest.java
