+++ src/test/java/net/sf/jabref/logic/search/GrammarBasedSearchRuleDescriberTest.java
