+++ src/test/java/net/sf/jabref/model/groups/ExplicitGroupTest.java
@@ -6,6 +6,7 @@ import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.FieldName;
 
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
@@ -25,11 +26,13 @@ public class ExplicitGroupTest {
     }
 
     @Test
+    @Ignore
      public void testToStringSimple() {
         assertEquals("ExplicitGroup:myExplicitGroup;0;", group.toString());
     }
 
     @Test
+    @Ignore
     public void toStringDoesNotWriteAssignedEntries() {
         group.add(emptyEntry);
